Rename building to install, --root arg

+++ install/module.py
@@ -46,14 +46,14 @@ import shutil
 def _key(path) -> int:
     # sorted from desired to undesired
 
-    if path.startswith('/usr/lib/python3/') or path == '/usr/lib/python3':
+    if path.startswith("/usr/lib/python3/") or path == "/usr/lib/python3":
         # Paths that work independent of the python version, yes please
         return -1
 
-    if path.startswith('/usr/local'):
+    if path.startswith("/usr/local"):
         return 1
 
-    if not path.startswith('/usr'):
+    if not path.startswith("/usr"):
         # Editable package paths, not system-wide, user installations which don't work
         # for input-remapper.
         return 2
@@ -74,32 +74,36 @@ def _get_packages_dir():
 
 
 def _get_commit_hash():
-    git_call = subprocess.check_output(['git', 'rev-parse', 'HEAD'])
+    git_call = subprocess.check_output(["git", "rev-parse", "HEAD"])
     commit = git_call.decode().strip()
     return commit
 
 
 def _fill_templates(target: str):
-    path = os.path.join(target, 'inputremapper', 'installation_info.py')
+    path = os.path.join(target, "inputremapper", "installation_info.py")
     print("Writing", path)
     assert os.path.exists(path)
-    with open(path, 'w') as f:
+    with open(path, "w") as f:
         lines = [
             f"COMMIT_HASH = '{_get_commit_hash()}'",
             "VERSION = '2.2.0'",
             "DATA_DIR = '/usr/share/inputremapper'",
         ]
 
-        f.write('\n'.join(lines))
+        f.write("\n".join(lines))
 
 
-def build_input_remapper_module():
+def build_input_remapper_module(root: str):
     # I'd use --prefix and --root, but
     # `pip install . --root ./build --prefix usr`
     # makes it end up in ./build/usr/local
 
-    # get_packages_dir is an absolute path, so os.path.join can't be used.
-    target = f"./build{_get_packages_dir()}"
+    package_dir = _get_packages_dir()
+    if package_dir.startswith("/"):
+        package_dir = package_dir[1:]
+
+    target = os.path.join(root, package_dir)
+
     subprocess.check_call(
         [
             sys.executable,
@@ -109,11 +113,13 @@ def build_input_remapper_module():
             ".",
             "--target",
             target,
-            "--no-deps"
+            "--no-deps",
         ]
     )
 
-    # pip puts its own leftovers into ./build that we don't need.'
-    shutil.rmtree('./build/lib/')
+    # pip puts its own leftovers into ./build that we don't need.
+    # This only happens, when root is set to "build".
+    if root.endswith("build") and os.path.exists("./build/lib/"):
+        shutil.rmtree("./build/lib/")
 
     _fill_templates(target)
